Move sources into gradle file structure
