--- WorldRestoration/src/as/minecraft/worldrestoration/utils/Utils.java
@@ -1,12 +1,0 @@
-package as.minecraft.worldrestoration.utils;
-
-import org.bukkit.ChatColor;
-
-public class Utils {
-	
-	//Translate standard Bukkit "&" color coding to actual colors
-	public static String chat (String s) {
-		return ChatColor.translateAlternateColorCodes('&', s);
-	}
-}
-
